Update to v091r11 release.
byuu says:

This release refines HSU1 support as a bidirectional protocol, nests SFC
manifests as "release/cartridge" and "release/information" (but release/
is not guaranteed to be finalized just yet), removes the database
integration, and adds support for ananke.

ananke represents inevitability. It's a library that, when installed,
higan can use to load files from the command-line, and also from a new
File -> Load Game menu option.

I need to change the build rules a bit for it to work on Windows (need
to make phoenix a DLL, basically), but it works now on Linux.

Right now, it only takes *.sfc file names, looks them up in the included
database, converts them to game folders, and returns the game folder
path for higan to load.

The idea is to continue expanding it to support everything we can that
I don't want in the higan core:
- load *.sfc, *.smc, *.swc, *.fig files
- remove SNES copier headers
- split apart merged firmware files
- pull in external firmware files (eg dsp1b.rom - these are staying
  merged, just as SPC7110 prg+dat are merged)
- load *.zip and *.7z archives
- prompt for selection on multi-file archives
- generate manifest files based on heuristics
- apply BPS patches

The "Load" menu option has been renamed to "Library", to represent games
in your library. I'm going to add some sort of suffix to indicate
unverified games, and use a different folder icon for those (eg
manifests built on heuristics rather than from the database.)

So basically, to future end users:
File -> Load Game will be how they play games.
Library -> (specific system) can be thought of as an infinitely-sized
    recent games list.

purify will likely become a simple stub that invokes ananke's functions.
No reason to duplicate all that code.

+++ higan/Makefile
@@ -34,8 +34,7 @@ endif
 # platform
 ifeq ($(platform),x)
   flags += -march=native
-  link += -ldl -lX11 -lXext
-else ifeq ($(platform),osx)
+  link += -Wl,-export-dynamic -ldl -lX11 -lXext
 else ifeq ($(platform),win)
   ifeq ($(arch),win32)
     flags += -m32
@@ -82,13 +81,14 @@ clean:
 	-@$(call delete,obj/*.dylib)
 	-@$(call delete,obj/*.dll)
 	-@$(call delete,*.res)
-	-@$(call delete,*.pgd)
-	-@$(call delete,*.pgc)
-	-@$(call delete,*.ilk)
-	-@$(call delete,*.pdb)
 	-@$(call delete,*.manifest)
 
+archive:
+	if [ -f higan.tar.xz ]; then rm higan.tar.xz; fi
+	tar -cJf higan.tar.xz `ls`
+
 sync:
+ifeq ($(shell id -un),byuu)
 	if [ -d ./libco ]; then rm -r ./libco; fi
 	if [ -d ./nall ]; then rm -r ./nall; fi
 	if [ -d ./ruby ]; then rm -r ./ruby; fi
@@ -103,9 +103,6 @@ sync:
 	rm -r ruby/_test
 	rm -r phoenix/nall
 	rm -r phoenix/test
-
-archive:
-	if [ -f higan.tar.xz ]; then rm higan.tar.xz; fi
-	tar -cJf higan.tar.xz `ls`
+endif
 
 help:;
